+++ controller/form_api_controller.js
@@ -17,7 +17,7 @@ const FormApi = Function.inherits('Alchemy.Controller', 'FormApi');
  *
  * @author   Jelle De Loecker   <jelle@elevenways.be>
  * @since    0.1.0
- * @version  0.1.0
+ * @version  0.1.5
  *
  * @param    {Conduit}   conduit
  */
@@ -26,8 +26,25 @@ FormApi.setAction(async function related(conduit) {
 	const body = conduit.body;
 
 	const model = this.getModel(body.assoc_model);
+	let crit = model.find();
+	crit.limit(50);
 
-	let records = await model.find('all');
+	if (body.config && body.config.search) {
+		let display_fields = Array.cast(model.displayField);
+
+		let or = crit.or();
+		let rx = RegExp.interpretWildcard('*' + body.config.search + '*', 'i');
+
+		for (let field of display_fields) {
+			if (!field) {
+				continue;
+			}
+
+			crit.where(field).equals(rx);
+		}
+	}
+
+	let records = await model.find('all', crit);
 
 	let result = {
 		available : records.available,